+++ src/com/mallen/flightui/ui/modules/NumberIndicatorHorizontalTape.java
@@ -15,15 +15,11 @@
 
 package com.mallen.flightui.ui.modules;
 
-import java.awt.Color;
 import java.awt.Font;
 import java.awt.FontMetrics;
 import java.awt.Graphics;
 import java.awt.geom.Rectangle2D;
 import java.awt.image.ImageObserver;
-
-import com.mallen.flightui.wrapper.FLUI_GLOBAL;
-import com.mallen.flightui.wrapper.flui.FLUIAircraft;
 
 public class NumberIndicatorHorizontalTape {
 	int x = 0, y = 0;
@@ -82,7 +78,7 @@ public class NumberIndicatorHorizontalTape {
 		int valueRef = (int) (value/valueMult);
 		
 		for(int i = -tapeRange/2; i < tapeRange/2; i++){
-			String s = "" + Math.round(((value/ ((int) (100* (double) valueMult))*(100*valueMult)))-((-100*valueMult)*i));
+			String s = "" + Math.round(((value/ ((int) (100* valueMult))*(100*valueMult)))-((-100*valueMult)*i));
 			if(s.length() < 3){
 				s = "0" + s;
 			}
